requirements.txt
awsebcli==3.14.10
blessed==1.15.0
bootstrap-admin==0.4.3
botocore==1.12.84
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.1.5
django-filter==2.1.0
django-grappelli==2.12.2
django-heroku==0.3.1
django-localflavor==2.1
django-mongodb-engine==0.6.0
djangorestframework==3.9.1
djangotoolbox==1.8.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.7
jmespath==0.9.3
jsonschema==2.6.0
Markdown==3.0.1
pathspec==0.5.5
Pillow==5.4.1
psycopg2==2.7.6.1
pymongo==3.2.2
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
virtualenv==16.2.0
wcwidth==0.1.7
websocket-client==0.54.0
whitenoise==4.1.2
yolk3k==0.9

